[data-drive-json-consumption] - Updated WorkflowJSONSpec access level to be internal, fixed a file_types_order issue - TT RAG
